% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-backend.R
\name{benchmark_parallel}
\alias{benchmark_parallel}
\title{Run benchmark parallely}
\usage{
benchmark_parallel(
  method,
  probnum,
  dims,
  rep,
  cec = 17,
  suite = "basic",
  cpupc = 0.75,
  write_flag = TRUE,
  benchmark_id,
  dest
)
}
\arguments{
\item{method}{optimization algorithm :: function}

\item{probnum}{indices of problem function :: [Int]}

\item{dims}{dimensionalities :: [Int]}

\item{rep}{amount of repetition :: Int}

\item{cec}{year of benchmark :: Int}

\item{suite}{benchmark suite :: String}

\item{cpupc}{CPU usage in pct :: Int}

\item{benchmark_id}{benchmark name :: String}

\item{dest}{filepath to place where benchmark restuls will be saved :: String}
}
\description{
Function runs benchmark `.cec` on `.method` for given test functions and dimensionality. 
Evalution on each function is repeated `.rep` times. User is able to
specify usage of CPU cores by `.cpupc` arg.
}
