% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cec_class_grid}
\alias{cec_class_grid}
\title{ECDF in function classes}
\usage{
cec_class_grid(filepaths, dim, cec = 17, rep = 51)
}
\arguments{
\item{filepaths}{filepaths to dirs with benchmark results :: [String]}

\item{dim}{dimensionality :: String}

\item{cec}{version of CEC :: Int}

\item{rep}{number of repetitions :: Int}
}
\description{
ECDF in function classes
}
