% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cec_problem_grid}
\alias{cec_problem_grid}
\title{ECDF per problem}
\usage{
cec_problem_grid(filepaths, cec, config)
}
\arguments{
\item{filepaths}{filepaths to dirs with benchmark results :: [String]}

\item{cec}{version of CEC :: Int}

\item{config}{map with benchmark config: :: [dim :: integer, probnums :: [integer], reps :: integer]}
}
\description{
ECDF per problem
}
