% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf-backend.R
\name{check_format}
\alias{check_format}
\title{Check file format}
\usage{
check_format(filepath)
}
\arguments{
\item{filepath}{:: [character]}
}
\value{
type of format :: [character]
}
\description{
Function checks if given `filepath` has JSON or TXT format.
}
