% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf-backend.R
\name{compute_aoc}
\alias{compute_aoc}
\title{Area under curve}
\usage{
compute_aoc(dfx, method = "Method", xarg = "Bstep", yarg = "Value")
}
\arguments{
\item{dfx}{data frame with ECDF values}

\item{method}{column name with algorithm(s) label(s) :: character}

\item{xarg}{column name of x-axis :: character}

\item{yarg}{column name of y-axis :: character}
}
\value{
data table with `Aoc` column :: tibble
}
\description{
Function computes area under ECDF curve (AOC)
}
