% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf-frontend.R
\name{ecdf_plot}
\alias{ecdf_plot}
\title{ECDF plot}
\usage{
ecdf_plot(dfx, aoc = FALSE, ymax = 0.3)
}
\arguments{
\item{dfx}{data frame with benchmark results}

\item{aoc}{binary flag}

\item{ymax}{height of the table with AOC values}
}
\description{
Function plots ECDF curves.
}
