% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_id}
\alias{extract_id}
\title{Extract id}
\usage{
extract_id(idpaths)
}
\arguments{
\item{idpaths}{list of filepaths to benchmark :: [character]}
}
\value{
list with raw ids of benchamrks :: [character]
}
\description{
Function extracts benchmarks ids from list of filepaths.
}
