% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_method}
\alias{extract_method}
\title{Extract method name}
\usage{
extract_method(id)
}
\arguments{
\item{id}{benchmark ID}
}
\description{
Function uses regex to extract method name from benchmark ID.
}
