% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_names}
\alias{extract_names}
\title{Extract algorithms names}
\usage{
extract_names(amount, algs)
}
\arguments{
\item{amount}{amount of benchmarked algorithms}

\item{algs}{list of closures}
}
\value{
list of algorithm names
}
\description{
Function extracts names of algorithm from given config file.
}
