% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf-backend.R
\name{generate_df}
\alias{generate_df}
\title{ECDF data frame}
\usage{
generate_df(
  .dim,
  .ids,
  .probnums,
  .source,
  .rep = 51,
  .bsteps = c(0.01, 0.02, 0.03, 0.05, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1) *
    log10(10000)
)
}
\arguments{
\item{.dim}{dimensionality of problem :: Int}

\item{.ids}{benchmark ids :: [String]}

\item{.probnums}{problem numbers :: [Int]}

\item{.source}{version of CEC benchmark :: Int}

\item{.rep}{number of repetition}

\item{.bsteps}{fraction of evaluation function budget}
}
\description{
Function generates data frame which is backend for ECDF plot function.
}
