% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf-backend.R
\name{generate_dfr}
\alias{generate_dfr}
\title{Compute data frame with ECDF}
\usage{
generate_dfr(idpaths, config, format_handler)
}
\arguments{
\item{idpaths}{list of paths :: [character]}

\item{config}{map with benchmark config: :: [dim :: integer, probnums :: [integer], reps :: integer]}

\item{format_handler}{function to handle specific format file :: integer -> [character] -> integer -> [[numeric]]}
}
\description{
Function for given `idpaths` and benchmark `config` generates data frame
with ECDF values.
}
