% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{generate_table}
\alias{generate_table}
\title{Compute data frame with ECDF}
\usage{
generate_table(idpaths, format_handler, probnums, dim)
}
\arguments{
\item{idpaths}{list of paths :: [character]}

\item{format_handler}{function to handle specific format file :: integer -> [character] -> integer -> [[numeric]]}

\item{probnums}{problem indices :: [numeric]}

\item{dim}{dimensionality of function :: numeric}
}
\description{
Function for given `idpaths` and benchmark setting generates data frame
with ECDF values using adequate format handler.
}
