% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_eval_func}
\alias{get_eval_func}
\title{Get function wrapper}
\usage{
get_eval_func(cec, suite)
}
\arguments{
\item{cec}{version of CEC (13, 14, 17, 21) :: Int}

\item{suite}{benchmark suite :: String}
}
\description{
Get function wrapper
}
