% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf.R
\name{get_mincnt}
\alias{get_mincnt}
\title{ugly af and needs refactoring
TODO}
\usage{
get_mincnt(.methods, .results, .ecdf, .probnums, .bsteps, .rep, .max_succ)
}
\description{
ugly af and needs refactoring
TODO
}
