% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_recordedTimes}
\alias{get_recordedTimes}
\title{Get budget steps for CEC2021}
\usage{
get_recordedTimes(dim)
}
\arguments{
\item{dim}{dimenstionality :: Int}
}
\description{
Get budget steps for CEC2021
}
