% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf-backend.R
\name{get_result}
\alias{get_result}
\title{Load data}
\usage{
get_result(.probnum, .ids, .dim, .cec)
}
\arguments{
\item{.probnum}{problem number :: [Int]}

\item{.ids}{benchmark ids :: [String]}

\item{.dim}{dimensionality of problem :: Int}

\item{.cec}{version of CEC benchmark :: Int}
}
\description{
Function reads results of given benchmark from file.
}
