% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_scores}
\alias{get_scores}
\title{Get values of global optimum}
\usage{
get_scores(cec, suite)
}
\arguments{
\item{cec}{version of CEC :: Int}

\item{suite}{benchmark suite :: String}
}
\description{
Get values of global optimum
}
