% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-utils.R
\name{\%<~\%}
\alias{\%<~\%}
\title{Alias for {zeallot} destructuring operator}
\usage{
x \%<~\% value
}
\description{
Operator %<-% is also used by {future} package and using operator with
package prefix i.e. {package}::%...% seems to be cumbersome.
}
\examples{
temp_list = list(a = 0, b = 1)
c(x, y) \%<~\% temp_list
}
