% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%++\%}
\alias{\%++\%}
\title{String concatenation operator}
\usage{
str1 \%++\% str2
}
\arguments{
\item{str1}{lhs string}

\item{str2}{rhs string}
}
\description{
Infix string concatenation operator
}
\examples{
"Hello" \%++\% "world"
}
