% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf-backend.R
\name{load_result_txt}
\alias{load_result_txt}
\title{Load data in TXT format}
\usage{
load_result_txt(probnum, idpaths, dim)
}
\arguments{
\item{probnum}{problem number :: integer}

\item{idpaths}{benchmark ids :: [character]}

\item{dim}{dimensionality of problem :: integer}
}
\description{
Function reads results of given benchmark from TXT file.
}
