% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf-backend.R
\name{parse_json}
\alias{parse_json}
\title{Get benchmark data from JSON}
\usage{
parse_json(json, probnum, dim)
}
\arguments{
\item{json}{parsed JSON :: list}

\item{probnum}{number of function in benchmark :: integer}

\item{dim}{dimensionality of function :: integer}
}
\description{
Function gets benchmark results from previously parsed JSON.
}
