% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{parse_yaml_config}
\alias{parse_yaml_config}
\title{YAML config parser}
\usage{
parse_yaml_config(filename)
}
\arguments{
\item{filename}{name of config file :: String}
}
\description{
Function parses YAML configuration file.
}
