% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{print_stats}
\alias{print_stats}
\title{Benchmark results basic stats}
\usage{
print_stats(vec)
}
\arguments{
\item{vec}{vector with results}
}
\description{
Function prints on STDIN basic statistics of benchmark result vector.
}
