% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-frontend.R
\name{run_benchmark}
\alias{run_benchmark}
\title{Benchmark frontend}
\usage{
run_benchmark(config)
}
\arguments{
\item{config}{a list written by user or read from YAML configuration file.}
}
\description{
Function takes configuration of benchmark list and runs benchmark.
}
