% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_eps}
\alias{save_eps}
\title{Save plot}
\usage{
save_eps(.plot, .name, .x = 6, .y = 6)
}
\arguments{
\item{.plot}{ggplot2 object}

\item{.name}{name of pllot}

\item{.x}{width of image}

\item{.y}{height of image}
}
\description{
Function saves ECDF plot to postcript.
}
