% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_metadata}
\alias{save_metadata}
\title{Save metadata}
\usage{
save_metadata(dest, filename, info)
}
\arguments{
\item{dest}{path to directory}

\item{filename}{name of file with metadata}

\item{info}{list with stored objects}
}
\value{
IO :: ()
}
\description{
Function save serialized benchmark's metadata. Created RDS file contains:
- benchmarked method written as a closure
- time of benchmark execution in minutes
- id of benchmark.
}
