% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_results}
\alias{save_results}
\title{Save benchmark results}
\usage{
save_results(x, dest, filename)
}
\arguments{
\item{x}{result vector or matrix}

\item{dest}{path to directory}

\item{filename}{name of json file}
}
\value{
IO :: ()
}
\description{
Function saves result of benchmark to JSON file. If given dirpath doesn't exist function creates it.
}
