% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-backend.R
\name{send_sms}
\alias{send_sms}
\title{Send SMS}
\usage{
send_sms(filepath, type, id)
}
\arguments{
\item{filepath}{path to Twilio auth configuration :: String}

\item{type}{type of message i.e 'start' or 'end' of benchmark :: String}

\item{id}{benchmark id :: String}
}
\description{
Function sends SMS with information about status of benchmark.
It reads number and Twilio auth from .twilio-meta file.
}
