% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf-backend.R
\name{split_formats}
\alias{split_formats}
\title{Check file format}
\usage{
split_formats(filepaths)
}
\arguments{
\item{filepath}{:: [character]}
}
\value{
type of format :: [character]
}
\description{
Function checks if given `filepath` has JSON or TXT format.
}
