% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{verify_config_names}
\alias{verify_config_names}
\title{Verify config list}
\usage{
verify_config_names(config)
}
\arguments{
\item{config}{named list :: [character]}
}
\value{
binary flag :: logical
}
\description{
Function checks if given config has only required fields.
}
