% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{get_dfr}
\alias{get_dfr}
\title{Benchmark results data frame}
\usage{
get_dfr(idpaths, config)
}
\arguments{
\item{idpaths}{filepath to benchmark data :: [character]}

\item{config}{map with benchmark config: :: [dim :: integer, probnums :: [integer], reps :: integer]}
}
\value{
data frame with benchmark data :: tibble
}
\description{
Function generates ready to plot data frame with benmchark results.
}
