% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-backend.R
\name{save_results}
\alias{save_results}
\title{Save benchmark results}
\usage{
save_results(x, cec, id, prob, dim, label, type)
}
\arguments{
\item{x}{result vector or matrix}

\item{cec}{CEC version :: Int}

\item{id}{benchmark id :: String}

\item{prob}{problem number :: Int}

\item{dim}{dimension of given problem :: Int}

\item{label}{label of algorithm :: String}

\item{type}{result type :: String}
}
\description{
Function saves result of benchmark to text file.
}
